requirements.txt
# Sai Subodh — pin versions locally if you need an exact repro.
flask>=2.0
werkzeug>=2.0
python-dotenv>=1.0
pandas>=1.3
numpy>=1.21
scikit-learn>=1.0
xgboost>=1.5
nltk>=3.6
matplotlib>=3.5
seaborn>=0.11
joblib>=1.1
sklearn-genetic>=0.2
